changes according to code review
