+++ apps/web/components/blocks/BannerCarousel/Banner.vue
@@ -17,7 +17,7 @@
       :class="['absolute inset-0 p-4 flex flex-col md:basis-2/4', { 'md:p-10': banner.text.bgcolor }]"
       :style="{
         color: banner.text.color,
-        textAlign: getTextAlignment(),
+        textAlign: getTextAlignment(banner.text?.textAlignment ?? ''),
         alignItems: getContentPosition(banner.text.align ?? ''),
         justifyContent: getContentPosition(banner.text.justify ?? ''),
       }"
@@ -99,24 +99,7 @@ const isMobile = computed(() => viewport.isLessThan('lg'));
 const props = defineProps<BannerProps & { slideIndex?: number }>();
 
 const banner = computed(() => props.content);
-
-const hexToRgba = (hex: string = '#fff', opacity: number = 1) => {
-  const cleanHex = hex.replace('#', '');
-
-  const fullHex =
-    cleanHex.length === 3
-      ? cleanHex
-          .split('')
-          .map((c) => c + c)
-          .join('')
-      : cleanHex;
-
-  const red = parseInt(fullHex.substring(0, 2), 16);
-  const green = parseInt(fullHex.substring(2, 4), 16);
-  const blue = parseInt(fullHex.substring(4, 6), 16);
-
-  return `rgba(${red}, ${green}, ${blue}, ${opacity})`;
-};
+const { hexToRgba, getImageHeight, getTextAlignment, getContentPosition } = useBlockContentHelper();
 
 const getImageUrl = () => {
   switch (viewport.breakpoint.value) {
@@ -131,61 +114,6 @@ const getImageUrl = () => {
     }
     default: {
       return banner.value.image?.mobile ?? '';
-    }
-  }
-};
-
-const getImageHeight = () => {
-  switch (viewport.breakpoint.value) {
-    case '4xl': {
-      return '768px';
-    }
-
-    case 'lg': {
-      return '576px';
-    }
-    case 'md': {
-      return '432px';
-    }
-    default: {
-      return '320px';
-    }
-  }
-};
-
-const getTextAlignment = () => {
-  const textAlignment = banner.value.text?.textAlignment ?? '';
-
-  switch (textAlignment) {
-    case 'center': {
-      return 'center';
-    }
-    case 'right': {
-      return 'right';
-    }
-    default: {
-      return 'left';
-    }
-  }
-};
-
-const getContentPosition = (axis: string) => {
-  if (isMobile.value) {
-    return 'center';
-  }
-
-  switch (axis) {
-    case 'center': {
-      return 'center';
-    }
-    case 'right': {
-      return 'flex-end';
-    }
-    case 'bottom': {
-      return 'flex-end';
-    }
-    default: {
-      return 'flex-start';
     }
   }
 };
