+++ apps/web/components/blocks/CategoryData/__tests__/CategoryData.spec.ts
@@ -34,13 +34,10 @@ const mockProps: CategoryDataProps = {
       paddingBottom: 0,
       paddingLeft: 0,
       paddingRight: 0,
+      narrowContainer: false,
     },
     displayCategoryImage: 'off',
     text: {
-      pretitle: 'Test pretitle',
-      title: 'Test title',
-      subtitle: 'Test subtitle',
-      htmlDescription: '<p>Test description</p>',
       bgColor: '#fff',
       bgOpacity: 1,
       background: true,
